requirements.txt
discord.py==2.5.2
pandas==2.2.3
dotenv==0.9.9